+++ components/create/aspectRatioButton.tsx
@@ -1,5 +1,8 @@
 import { connect } from "react-redux";
 
+import { store } from "@/store";
+import { cycleAspectRatio } from "@/store/actions";
+
 import Button from "@/components/shared/Button";
 
 type AspectRatioProps = {
@@ -9,7 +12,14 @@ type AspectRatioProps = {
 const AspectRatioButton = ({ ratio }: AspectRatioProps) => {
   const buttonLabel = ratio ? `${ratio.width}:${ratio.height}` : "AR";
 
-  return <Button label={buttonLabel} disabled={true}></Button>;
+  return (
+    <Button
+      label={buttonLabel}
+      onPress={() => {
+        store.dispatch(cycleAspectRatio());
+      }}
+    ></Button>
+  );
 };
 
 const mapStateToProps = (state: any) => {
@@ -18,4 +28,8 @@ const mapStateToProps = (state: any) => {
   };
 };
 
-export default connect(mapStateToProps)(AspectRatioButton);
+const mapDispatchToProps = {
+  cycleAspectRatio,
+};
+
+export default connect(mapStateToProps, mapDispatchToProps)(AspectRatioButton);