+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - An APP icon! [#51](https://github.com/pLTdwaxn/Caroushow/issues/51)
 - EAS project config [[#63](https://github.com/pLTdwaxn/Caroushow/issues/63)]
 - Prebuild iOS with Expo New Architecture [#82](https://github.com/pLTdwaxn/Caroushow/issues/82)
+- Ability to cycle through Aspect Ratios [#87](https://github.com/pLTdwaxn/Caroushow/issues/87)
 
 ### Changed